Create restrict_comments_to_authors_and_post_author
This code will restrict the visibility of comments to only the post author and the comment author. If the current user is neither the post author nor the comment author, they will see the text "This comment is private." instead of the actual comment. This function will also work for logged-out users, hiding all comments from them.

Please note that this code doesn't restrict visibility in the admin area, so administrators will still be able to see all comments.